back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "W1 Hackathon Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.4.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]


[tool.ruff]
line-length = 100
target-version = "py313"

fix = true
show-fixes = true
output-format = "grouped"

# Excluir diretórios comumente ignorados
extend-exclude = [
    "migrations/*"
]

# Seleção de regras - configuração extensiva baseada na documentação
lint.select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "UP",  # pyupgrade
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "RUF", # ruff-specific rules
    # "D",   # pydocstyle
    "ANN", # flake8-annotations
    # "S",   # flake8-bandit
    # "T",   # flake8-debugger
]


# Ignorar regras específicas que podem ser muito restritivas
lint.ignore = [
    "D107",  # Missing docstring in __init__
    "D203",  # 1 blank line required before class docstring
    "D213",  # Multi-line docstring summary should start at the second line
    "ANN001", # Missing type annotation for self in method
    "ANN002",  # Missing type annotation for cls in classmethod
    "ANN003"  # Missing type annotation for kwargs
]

# Ativar fixação automática para todas as regras selecionadas
lint.fixable = ["ALL"]
lint.unfixable = []

lint.isort.known-first-party = ["olim"]
lint.isort.section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lint.isort.combine-as-imports = true

lint.pydocstyle.convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

